fix(ui): Display better summary for arbitrary time ranges
